pyproject.toml
[tool.poetry]
name = "openapi_client_generator"
version = "0.0.1"
description = ""
authors = ["sasano8 <y-sasahara@ys-method.com>"]
license = "MIT"
readme = "README.md"
# repository = "https://github.com/sasano8/asy"

[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.18.2"
pydantic = "^1.8.2"
asgi-lifespan = "1.*"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"
mypy = "^0.812"
pytest = "^6.2.2"
pre-commit = "^2.12.0"
devtools = "^0.6.1"
fastapi = "^0.65.1"
uvicorn = "^0.14.0"
sphinx = "^4.0.2"
sphinx_rtd_theme = "^0.5.2"
recommonmark = "^0.7.1"
pyyaml = "^5.4.1"
pytest-asyncio = "^0.15.1"
typer = "^0.3.2"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
aiofiles = "^0.7.0"
isort = "^5.9.3"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
